ТЗ 15  fix bag
